# lib/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Agendamento v2",
    "version": "1.0.0",
    "description": "API REST completa para sistema de agendamentos com autenticação via API Key",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key no formato: sk_<apiClientId>_<random>"
      },
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token do Supabase (apenas para endpoints administrativos)"
      }
    },
    "schemas": {
...
      "Availability": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "professionalId": {
            "type": "string",
            "format": "uuid"
          },
          "dayOfWeek": {
            "type": "integer",
            "description": "0 = Domingo, 1 = Segunda, ..., 6 = Sábado",
            "minimum": 0,
            "maximum": 6
          },
          "startTime": {
            "type": "string",
            "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
          },
          "endTime": {
            "type": "string",
            "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/availabilities": {
      "get": {
        "tags": ["Agendamentos"],
        "summary": "Listar Availabilities",
        "description": "Lista todas as availabilities da company associada à API Key. Pode filtrar por professionalId.\n\n**Exemplo de requisição:**\n```bash\ncurl -X GET \"http://localhost:3000/api/v1/availabilities?professionalId=123e4567-e89b-12d3-a456-426614174000\" \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\"\n```",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {
            "name": "professionalId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do professional (opcional - filtra por professional)",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de availabilities retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Availability"}
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "professionalId": "123e4567-e89b-12d3-a456-426614174001",
                      "dayOfWeek": 1,
                      "startTime": "09:00",
                      "endTime": "18:00",
                      "created_at": "2024-01-01T00:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Agendamentos"],
        "summary": "Criar Availability",
        "description": "Cria uma nova disponibilidade para um professional.\n\n**Exemplo de requisição:**\n```bash\ncurl -X POST http://localhost:3000/api/v1/availabilities \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"professionalId\": \"123e4567-e89b-12d3-a456-426614174000\",\n    \"dayOfWeek\": 1,\n    \"startTime\": \"09:00\",\n    \"endTime\": \"18:00\"\n  }'\n```",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["professionalId", "dayOfWeek", "startTime", "endTime"],
                "properties": {
                  "professionalId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  },
                  "dayOfWeek": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 6,
                    "description": "0 = Domingo, 1 = Segunda, ..., 6 = Sábado",
                    "example": 1
                  },
                  "startTime": {
                    "type": "string",
                    "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$",
                    "example": "09:00"
                  },
                  "endTime": {
                    "type": "string",
                    "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$",
                    "example": "18:00"
                  }
                }
              },
              "example": {
                "professionalId": "123e4567-e89b-12d3-a456-426614174000",
                "dayOfWeek": 1,
                "startTime": "09:00",
                "endTime": "18:00"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Availability criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/Availability"}
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "professionalId": "123e4567-e89b-12d3-a456-426614174001",
                    "dayOfWeek": 1,
                    "startTime": "09:00",
                    "endTime": "18:00",
                    "created_at": "2024-01-01T00:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Validation error",
                  "errors": {
                    "dayOfWeek": ["Number must be less than or equal to 6"],
                    "startTime": ["String must match pattern"]
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "404": {
            "description": "Professional não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Professional not found or doesn't belong to your company"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}